requirements.txt
setuptools
Django~=2.0.7
mysqlclient
Pillow
django-recaptcha
gunicorn==19.9.0
djangorestframework
django-filter
markdown